AOC-24 Starting day 14

+++ Days/Day13/InputParser.cs
@@ -25,7 +25,7 @@ public class InputParser
         }
     }
 
-    public int SumTokenCost()
+    public long SumTokenCost()
     {
         return this.clawMachines.Sum(claw => claw.GetMinCoinCount());
     }
